chore(python): remove vendored sympy, inline igcdex into starkware math_utils
The vendored sympy package was only used for the igcdex extended
Euclidean helper consumed by starkware/crypto/math_utils.py. Inline a
pure-python igcdex (identical results to the sympy python ground-types
implementation) and drop the whole static_dependencies/sympy tree.
